nmap_scan/requirements.txt
netaddr==0.7.19
python-nmap==0.6.1
